server/requirements.txt
Flask
    python-dotenv
    SQLAlchemy
    Werkzeug
    psycopg2-binary
    openai
    requests
    Flask-Cors
    mercadopago
    PyJWT
